server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai==1.12.0
numpy==1.26.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2==2.9.9
requests==2.31.0
